moved alert component
