+++ layouts/ScreenLayout.vue
@@ -9,7 +9,7 @@
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
-import Alert from '~/components/Alert.vue';
+import Alert from '~/components/common/Alert.vue';
 
 @Component({
   components: { Alert },
